refactor: 트리뷰 파일을 페이지 폴더 내부로 이동
