+++ apps/client/src/pages/structure/components/index.ts
