+++ apps/client/src/pages/structure/components/tree-memo-base-node/tree-memo-base-node.tsx
