+++ apps/client/src/pages/structure/components/tree-memo-list-node/components/detail-modal/detail-modal.tsx
