+++ apps/client/src/pages/structure/index.ts
@@ -1,0 +1,1 @@
+export { default as StructurePage } from './structure-page';
