+++ apps/client/src/pages/structure/structure-page.css.ts
