+++ apps/client/src/pages/structure/structure-page.tsx
@@ -1,10 +1,67 @@
-import TreeView from '@shared/components/memo-list-view/components/tree-view/tree-view';
+import {
+  BuiltInEdge,
+  Controls,
+  DefaultEdgeOptions,
+  EdgeTypes,
+  NodeTypes,
+  ReactFlow,
+} from '@xyflow/react';
+
+import { themeVars } from '@cds/ui';
+
+import { useReadMemoStructure } from './apis/queries';
+import { CustomEdge, TreeMemoBaseNode, TreeMemoListNode } from './components';
+import { buildTreeGraph } from './utils/build-tree-graph';
+
+import '@xyflow/react/dist/style.css';
+import * as styles from './structure-page.css';
+
+const ZOOM = {
+  MIN: 0.5,
+  MAX: 0.9,
+};
+
+const nodeTypes: NodeTypes = {
+  treeMemo: TreeMemoListNode,
+  baseMemo: TreeMemoBaseNode,
+};
+const edgeTypes: EdgeTypes = {
+  customEdge: CustomEdge,
+};
+const defaultEdgeOptions: DefaultEdgeOptions | Partial<BuiltInEdge> = {
+  type: 'smoothstep',
+  style: { strokeWidth: '0.1rem', stroke: themeVars.color.grey400 },
+  pathOptions: { borderRadius: 30 },
+};
 
 const StructurePage = () => {
+  const { data: memos, isPending } = useReadMemoStructure();
+  if (isPending || !memos) return null;
+
+  const { nodes, edges } = buildTreeGraph(memos);
+
   return (
-    <>
-      <TreeView />
-    </>
+    <div className={styles.container}>
+      <ReactFlow
+        nodes={nodes}
+        edges={edges}
+        nodeTypes={nodeTypes}
+        defaultEdgeOptions={defaultEdgeOptions}
+        edgeTypes={edgeTypes}
+        fitView
+        fitViewOptions={{ padding: 0.2 }}
+        minZoom={ZOOM.MIN}
+        maxZoom={ZOOM.MAX}
+        nodesDraggable={false}
+        nodesConnectable={false}
+        onNodeClick={(e) => {
+          e.stopPropagation();
+        }}
+        nodeDragThreshold={100}
+      >
+        <Controls />
+      </ReactFlow>
+    </div>
   );
 };
