+++ apps/client/src/pages/structure/utils/build-tree-graph.ts
